+++ icegraph/data/inspect/__init__.py
@@ -1,4 +1,6 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-__version__ = "0.10.0"
+from .inspect_lmdb import InspectLMDB
+
+__all__ = ["InspectLMDB"]